stepper updates
